requirements.txt
anthropic>=0.40.0
python-telegram-bot>=21.0
cairosvg>=2.7.0
python-dotenv>=1.0.0
